requirements.txt
# AIEMCS - Python Dependencies
# Install with: pip install -r requirements.txt

# ── Web Framework ──────────────────────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.30.1

# ── Database ───────────────────────────────────────────────────────────────────
sqlalchemy==2.0.30
pymysql==1.1.1
python-dotenv==1.0.1

# ── Authentication ─────────────────────────────────────────────────────────────
bcrypt==4.1.3
python-jose[cryptography]==3.3.0

# ── Validation ─────────────────────────────────────────────────────────────────
pydantic[email]==2.7.1

# ── File uploads ───────────────────────────────────────────────────────────────
python-multipart==0.0.9

# ── OCR (optional but recommended) ────────────────────────────────────────────
pytesseract==0.3.10
Pillow==10.3.0

# ── LangChain AI agent (optional) ─────────────────────────────────────────────
# Uncomment if you have an OpenAI API key:
# langchain==0.2.5
# langchain-community==0.2.5
# langchain-openai==0.1.8
# openai==1.30.3

# ── Seed data generation ───────────────────────────────────────────────────────
# (bcrypt already listed above)
